Update shape when changing vertex type. Fixes #57.
There's some weird behaviour with W-nodes if `W_INPUT` rather than `W_OUTPUT` is the selected node when its type is changed, and it's also not possible to change another node type into a W-node. This behaviour already exists, but the change introduced in this PR makes it more visually obvious.