Add comment to note quirk about qubit numbering of the demo graph.
The way the graph is initialised can be confusing to someone using the demo
graph to debug, since the qubit numbering used in the actual graph is shifted
by 4 from the numbering in the lists used to initialise it (due to the 4 added
input nodes).

See https://github.com/Quantomatic/zxlive/pull/213#issuecomment-1859038786 for context.